Deprecate nn.Container in favor of nn.Module
