+++ test/data/network2.py
@@ -1,7 +1,7 @@
 import torch.nn as nn
 
 
-class Net(nn.Container):
+class Net(nn.Module):
     def __init__(self):
         super(Net, self).__init__()
         self.linear = nn.Linear(10, 20)
